fix(place): split per-stadium sync into its own bean (self-invocation)
syncAll()이 같은 클래스 안의 syncForStadium()을 this로 직접 호출하는
self-invocation 구조라, syncForStadium에 붙은 @Transactional이 Spring AOP
프록시를 거치지 않아 무시되고 있었다(기존 restaurant 브랜치 코드 리뷰에서
지적된 부분).

@Transactional이 붙은 트랜잭션 단위(syncForStadium)를 PlaceStadiumSyncService
라는 별도 빈으로 분리했다. PlaceSyncService.syncAll()은 이제 다른 빈의
메서드를 호출하므로 프록시를 정상적으로 거치고, 트랜잭션이 실제로
적용된다. PlaceSyncServiceTest는 syncAll의 순회/예외격리 로직만 검증하도록
슬림화했고, 실제 동기화 로직(신규 저장/업데이트/삭제/재시도/상세정보 등)
검증은 PlaceStadiumSyncServiceTest로 옮겼다. 재시도가 모두 실패했을 때
기존 데이터가 삭제/저장되지 않는지 확인하는 회귀 테스트도 추가했다.
