+++ backend/app/src/test/java/com/yagubogu/place/service/PlaceSyncServiceTest.java
@@ -1,45 +1,33 @@
 package com.yagubogu.place.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyDouble;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.never;
+import static org.mockito.BDDMockito.willThrow;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.yagubogu.place.client.TourApiClient;
-import com.yagubogu.place.client.TourApiException;
-import com.yagubogu.place.domain.Place;
 import com.yagubogu.place.domain.PlaceCategory;
-import com.yagubogu.place.dto.PlaceParam;
-import com.yagubogu.place.repository.PlaceRepository;
 import com.yagubogu.stadium.domain.Stadium;
 import com.yagubogu.stadium.domain.StadiumLevel;
 import com.yagubogu.stadium.repository.StadiumRepository;
 import java.lang.reflect.Field;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+/**
+ * syncAll()의 오케스트레이션만 검증한다. 구장 하나에 대한 실제 동기화 로직은
+ * {@link PlaceStadiumSyncServiceTest}에서 다룬다 (트랜잭션 빈이 분리되어 있음).
+ */
 @ExtendWith(MockitoExtension.class)
 class PlaceSyncServiceTest {
 
     @Mock
-    private TourApiClient tourApiClient;
-
-    @Mock
-    private PlaceRepository placeRepository;
+    private PlaceStadiumSyncService placeStadiumSyncService;
 
     @Mock
     private StadiumRepository stadiumRepository;
@@ -47,175 +35,34 @@ class PlaceSyncServiceTest {
     @InjectMocks
     private PlaceSyncService placeSyncService;
 
-    private Stadium jamsil;
-    private Stadium suwon;
-
-    @BeforeEach
-    void setUp() {
-        jamsil = stadium(1L, "잠실", 37.5122, 127.0715);
-        suwon = stadium(2L, "수원", 37.2998, 127.0097);
-    }
-
-    @DisplayName("API에서 받은 신규 장소를 DB에 저장한다")
+    @DisplayName("syncAll은 모든 구장 x 카테고리 조합에 대해 동기화를 위임한다")
     @Test
-    void syncForStadium_신규_장소_저장() {
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("1001", 1L, "맛집A", "서울 송파구", 127.07, 37.51, 300, "02-111", null)
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT)).willReturn(List.of());
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        ArgumentCaptor<List<Place>> captor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).saveAll(captor.capture());
-        assertThat(captor.getValue()).hasSize(1);
-        assertThat(captor.getValue().get(0).getContentId()).isEqualTo("1001");
-        assertThat(captor.getValue().get(0).getTitle()).isEqualTo("맛집A");
-        assertThat(captor.getValue().get(0).getCategory()).isEqualTo(PlaceCategory.RESTAURANT);
-        verify(placeRepository, never()).deleteAll(any());
-    }
-
-    @DisplayName("이미 존재하는 장소의 정보를 업데이트한다")
-    @Test
-    void syncForStadium_기존_장소_업데이트() {
-        Place existing = Place.of(PlaceCategory.RESTAURANT, "1001", 1L, "구 이름", "구 주소",
-                127.0, 37.5, 500, null, null);
-
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("1001", 1L, "새 이름", "새 주소", 127.07, 37.51, 300, "02-111", "https://img.com")
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT))
-                .willReturn(List.of(existing));
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        ArgumentCaptor<List<Place>> captor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).saveAll(captor.capture());
-        Place updated = captor.getValue().get(0);
-        assertThat(updated.getTitle()).isEqualTo("새 이름");
-        assertThat(updated.getAddress()).isEqualTo("새 주소");
-        assertThat(updated.getDistance()).isEqualTo(300);
-        assertThat(updated.getImageUrl()).isEqualTo("https://img.com");
-    }
-
-    @DisplayName("API 응답에 없는 기존 장소는 삭제한다")
-    @Test
-    void syncForStadium_사라진_장소_삭제() {
-        Place stale = Place.of(PlaceCategory.RESTAURANT, "GONE", 1L, "없어진 맛집", null,
-                127.0, 37.5, null, null, null);
-
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("NEW", 1L, "새 맛집", null, 127.07, 37.51, 100, null, null)
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT))
-                .willReturn(List.of(stale));
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        ArgumentCaptor<List<Place>> deleteCaptor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).deleteAll(deleteCaptor.capture());
-        assertThat(deleteCaptor.getValue()).hasSize(1);
-        assertThat(deleteCaptor.getValue().get(0).getContentId()).isEqualTo("GONE");
-    }
-
-    @DisplayName("API 호출이 실패하면 재시도하고, 성공 시 동기화를 완료한다")
-    @Test
-    void syncForStadium_재시도_후_성공() {
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("1001", 1L, "맛집A", null, 127.07, 37.51, 100, null, null)
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willThrow(new TourApiException("timeout", new RuntimeException()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT)).willReturn(List.of());
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        verify(tourApiClient, times(2))
-                .fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble());
-        verify(placeRepository).saveAll(any());
-    }
-
-    @DisplayName("syncAll은 모든 구장 x 카테고리 조합에 대해 동기화하고, 한 조합이 실패해도 나머지는 계속한다")
-    @Test
-    void syncAll_모든_구장_카테고리_조합_진행() {
+    void syncAll_모든_구장_카테고리_조합에_대해_위임() {
+        Stadium jamsil = stadium(1L, "잠실", 37.5122, 127.0715);
+        Stadium suwon = stadium(2L, "수원", 37.2998, 127.0097);
         given(stadiumRepository.findAll()).willReturn(List.of(jamsil, suwon));
-        given(tourApiClient.fetchPlacesNear(any(PlaceCategory.class), anyLong(), anyDouble(), anyDouble()))
-                .willThrow(new TourApiException("fail", new RuntimeException()));
 
         placeSyncService.syncAll();
 
-        // 2개 구장 x 5개 카테고리 x 최대 3회 재시도 = 30번 호출
-        verify(tourApiClient, times(30))
-                .fetchPlacesNear(any(PlaceCategory.class), anyLong(), anyDouble(), anyDouble());
-        verify(placeRepository, never()).saveAll(any());
+        // 2개 구장 x 5개 카테고리 = 10번 위임
+        verify(placeStadiumSyncService, times(10))
+                .syncForStadium(any(Stadium.class), any(PlaceCategory.class));
     }
 
-    @DisplayName("API 응답이 비어있으면 기존 데이터를 모두 삭제한다")
+    @DisplayName("한 구장/카테고리 조합에서 동기화가 실패해도 나머지 조합은 계속 진행한다")
     @Test
-    void syncForStadium_API_빈_응답_기존_데이터_삭제() {
-        Place existing = Place.of(PlaceCategory.RESTAURANT, "OLD", 1L, "기존 맛집", null,
-                127.0, 37.5, null, null, null);
+    void syncAll_한_조합_실패해도_나머지_계속_진행() {
+        Stadium jamsil = stadium(1L, "잠실", 37.5122, 127.0715);
+        Stadium suwon = stadium(2L, "수원", 37.2998, 127.0097);
+        given(stadiumRepository.findAll()).willReturn(List.of(jamsil, suwon));
+        willThrow(new RuntimeException("sync failed"))
+                .given(placeStadiumSyncService).syncForStadium(any(Stadium.class), any(PlaceCategory.class));
 
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(List.of());
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT))
-                .willReturn(List.of(existing));
+        placeSyncService.syncAll();
 
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        ArgumentCaptor<List<Place>> saveCaptor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).saveAll(saveCaptor.capture());
-        assertThat(saveCaptor.getValue()).isEmpty();
-
-        ArgumentCaptor<List<Place>> deleteCaptor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).deleteAll(deleteCaptor.capture());
-        assertThat(deleteCaptor.getValue()).isNotEmpty();
-    }
-
-    @DisplayName("신규 장소는 상세 정보를 조회해 저장한다")
-    @Test
-    void syncForStadium_신규_장소_상세_정보_수집() {
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("1001", 1L, "맛집A", null, 127.07, 37.51, 100, null, null)
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT)).willReturn(List.of());
-        given(tourApiClient.fetchDetail(eq("1001"), eq(39))).willReturn("{\"common\":{}}");
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        verify(tourApiClient).fetchDetail("1001", 39);
-        ArgumentCaptor<List<Place>> captor = ArgumentCaptor.forClass(List.class);
-        verify(placeRepository).saveAll(captor.capture());
-        assertThat(captor.getValue().get(0).getDetailInfo()).isEqualTo("{\"common\":{}}");
-    }
-
-    @DisplayName("이미 상세 정보가 있는 장소는 다시 조회하지 않는다")
-    @Test
-    void syncForStadium_기존_상세_정보_있으면_재조회_안함() {
-        Place existing = Place.of(PlaceCategory.RESTAURANT, "1001", 1L, "구 이름", "구 주소",
-                127.0, 37.5, 500, null, null);
-        existing.updateDetailInfo("{\"common\":{\"overview\":\"기존 정보\"}}");
-
-        List<PlaceParam> fetched = List.of(
-                new PlaceParam("1001", 1L, "새 이름", "새 주소", 127.07, 37.51, 300, "02-111", null)
-        );
-        given(tourApiClient.fetchPlacesNear(eq(PlaceCategory.RESTAURANT), anyLong(), anyDouble(), anyDouble()))
-                .willReturn(fetched);
-        given(placeRepository.findAllByStadiumIdAndCategory(1L, PlaceCategory.RESTAURANT))
-                .willReturn(List.of(existing));
-
-        placeSyncService.syncForStadium(jamsil, PlaceCategory.RESTAURANT);
-
-        verify(tourApiClient, never()).fetchDetail(anyString(), anyInt());
+        // 예외가 나도 순회 자체는 10번(2 구장 x 5 카테고리) 다 시도된다
+        verify(placeStadiumSyncService, times(10))
+                .syncForStadium(any(Stadium.class), any(PlaceCategory.class));
     }
 
     // Stadium은 @GeneratedValue이므로 리플렉션으로 ID 주입
